moved predator prey and corridor into examples subdir
