+++ tests/test_turn_based_predator_prey.py
@@ -1,7 +1,7 @@
 
 import numpy as np
 
-from admiral.envs.predator_prey import PredatorPreyEnv, Predator, Prey
+from admiral.envs.examples.predator_prey import PredatorPreyEnv, Predator, Prey
 from admiral.managers import TurnBasedManager
 
 def test_turn_based_predator_prey_distance():
